Точно полностью выполненная домашняя работа
Поправила сообщения на ожидаемые из документации и фактические в flyUndefined и swimUnsuccessful + заменила функции создания уточки в тестах на новые версии с созданием строк
