+++ src/test/java/autotest/tests/DuckFlyTest.java
@@ -1,0 +1,102 @@
+package autotest.tests;
+
+import com.consol.citrus.TestCaseRunner;
+import com.consol.citrus.annotations.CitrusResource;
+import com.consol.citrus.annotations.CitrusTest;
+import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.testng.annotations.Optional;
+import org.testng.annotations.Test;
+
+import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
+import static com.consol.citrus.http.actions.HttpActionBuilder.http;
+
+public class DuckFlyTest extends TestNGCitrusSpringSupport {
+    @Test(description = "Полёт, ACTIVE крылья")
+    @CitrusTest
+    public void flyActive(@Optional @CitrusResource TestCaseRunner runner) {
+        String color = "yellow";
+        double height = 8.0;
+        String material = "rubber";
+        String sound = "quack";
+        String wingsState = "ACTIVE";
+        createDuck(runner, color, height, material, sound, wingsState);
+        runner.$(http().client("http://localhost:2222")
+                .receive()
+                .response()
+                .message()
+                .extract(fromBody().expression("$.id", "duckId"))
+        );
+        duckFly(runner, "${duckId}");
+        validateResponse(runner, "{\n" + "\"message\":\"I am flying\"\n" + "}");
+    }
+
+    @Test(description = "Полёт, FIXED крылья")
+    @CitrusTest
+    public void flyFixed(@Optional @CitrusResource TestCaseRunner runner) {
+        String color = "yellow";
+        double height = 8.0;
+        String material = "rubber";
+        String sound = "quack";
+        String wingsState = "FIXED";
+        createDuck(runner, color, height, material, sound, wingsState);
+        runner.$(http().client("http://localhost:2222")
+                .receive()
+                .response()
+                .message()
+                .extract(fromBody().expression("$.id", "duckId"))
+        );
+        duckFly(runner, "${duckId}");
+        validateResponse(runner, "{\n" + "\"message\":\"I can't fly\"\n" + "}");
+    }
+
+    @Test(description = "Полёт, UNDEFINED крылья")
+    @CitrusTest
+    public void flyUndefined(@Optional @CitrusResource TestCaseRunner runner) {
+        String color = "yellow";
+        double height = 8.0;
+        String material = "rubber";
+        String sound = "quack";
+        String wingsState = "UNDEFINED";
+        createDuck(runner, color, height, material, sound, wingsState);
+        runner.$(http().client("http://localhost:2222")
+                .receive()
+                .response()
+                .message()
+                .extract(fromBody().expression("$.id", "duckId"))
+        );
+        duckFly(runner, "${duckId}");
+        validateResponse(runner, "{\n" + "\"message\":\"Wings are not detected :(\"\n" + "}");
+    }
+    public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .post("api/duck/create")
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body("{" +
+                        "\"color\": \"" + color +
+                        "\",\n \"height\": \"" + height +
+                        "\",\n \"material\": \"" + material +
+                        "\",\n \"sound\": \"" + sound +
+                        "\", \n \"wingsState\": \"" + wingsState +
+                        "\" \n }"));
+    }
+
+    public void duckFly(TestCaseRunner runner, String id) { //лететь
+        runner.$(http().client("http://localhost:2222")
+                .send()
+                .get("api/duck/action/fly")
+                .queryParam("id", id));
+    }
+
+    public void validateResponse(TestCaseRunner runner, String responseMessage) {
+        runner.$(http().client("http://localhost:2222")
+                .receive()
+                .response(HttpStatus.OK)
+                .message()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(responseMessage));
+    }
+}
