+++ src/test/java/autotest/tests/DuckUpdateColorHeightTest.java
@@ -1,8 +1,9 @@
-package autotest;
+package autotest.tests;
 
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
+import com.consol.citrus.context.TestContext;
 import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -12,48 +13,25 @@ import org.testng.annotations.Test;
 import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
-public class DuckFlyTest extends TestNGCitrusSpringSupport {
-    @Test(description = "Полёт, ACTIVE крылья")
+public class DuckUpdateColorHeightTest extends TestNGCitrusSpringSupport {
+
+    @Test(description = "Обновление цвета и высоты уточки")
     @CitrusTest
-    public void flyActive(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 5.0, "rubber", "quack", "ACTIVE");
+    public void updateColorAndHeight(@Optional @CitrusResource TestCaseRunner runner, @Optional @CitrusResource TestContext context) {
+        String material = "rubber";
+        String sound = "quack";
+        String wingsState = "ACTIVE";
+        createDuck(runner, "yellow", 5.0, material, sound, wingsState);
         runner.$(http().client("http://localhost:2222")
                 .receive()
                 .response()
                 .message()
                 .extract(fromBody().expression("$.id", "duckId"))
         );
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{ \"message\": \"string\" }");
+        duckUpdate(runner, "white", 8.0, Integer.parseInt(context.getVariable("duckId")), material, sound, wingsState);
+        validateResponse(runner, "{ \"message\": \"Duck with id = " + "${duckId}" + " is updated\" }");
     }
 
-    @Test(description = "Полёт, FIXED крылья")
-    @CitrusTest
-    public void flyFixed(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 8.0, "rubber", "quack", "FIXED");
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response()
-                .message()
-                .extract(fromBody().expression("$.id", "duckId"))
-        );
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{ \"message\": \"string\" }");
-    }
-
-    @Test(description = "Полёт, UNDEFINED крылья")
-    @CitrusTest
-    public void flyUndefined(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 11.0, "rubber", "quack", "UNDEFINED");
-        runner.$(http().client("http://localhost:2222")
-                .receive()
-                .response()
-                .message()
-                .extract(fromBody().expression("$.id", "duckId"))
-        );
-        duckFly(runner, "${duckId}");
-        validateResponse(runner, "{\n" + "\"message\":\"Wings are not detected :(\"\n" + "}");
-    }
     public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
         runner.$(http().client("http://localhost:2222")
                 .send()
@@ -69,11 +47,17 @@ public class DuckFlyTest extends TestNGCitrusSpringSupport {
                         "\" \n }"));
     }
 
-    public void duckFly(TestCaseRunner runner, String id) { //лететь
+    public void duckUpdate(TestCaseRunner runner, String color, double height, int id, String material, String sound, String wingsState) {
         runner.$(http().client("http://localhost:2222")
                 .send()
-                .get("api/duck/action/fly")
-                .queryParam("id", id));
+                .put("api/duck/update")
+                .queryParam("color", color)
+                .queryParam("height", Double.toString(height))
+                .queryParam("id", Integer.toString(id))
+                .queryParam("material", material)
+                .queryParam("sound", sound)
+                .queryParam("wingsState", wingsState));
+
     }
 
     public void validateResponse(TestCaseRunner runner, String responseMessage) {
