+++ src/test/java/autotest/tests/DuckUpdateColorSoundTest.java
@@ -1,8 +1,9 @@
-package autotest;
+package autotest.tests;
 
 import com.consol.citrus.TestCaseRunner;
 import com.consol.citrus.annotations.CitrusResource;
 import com.consol.citrus.annotations.CitrusTest;
+import com.consol.citrus.context.TestContext;
 import com.consol.citrus.testng.spring.TestNGCitrusSpringSupport;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -12,21 +13,23 @@ import org.testng.annotations.Test;
 import static com.consol.citrus.dsl.MessageSupport.MessageBodySupport.fromBody;
 import static com.consol.citrus.http.actions.HttpActionBuilder.http;
 
-public class DuckDeleteTest extends TestNGCitrusSpringSupport {
+public class DuckUpdateColorSoundTest extends TestNGCitrusSpringSupport {
 
-
-    @Test(description = "РЈРґР°Р»РµРЅРёРµ СѓС‚РєРё")
+    @Test(description = "Обновление цвета и высоты уточки")
     @CitrusTest
-    public void deleteExisting(@Optional @CitrusResource TestCaseRunner runner) {
-        createDuck(runner, "yellow", 8.0, "rubber", "quack", "ACTIVE");
+    public void updateColorAndSound(@Optional @CitrusResource TestCaseRunner runner, @Optional @CitrusResource TestContext context) {
+        double height = 5.0;
+        String material = "rubber";
+        String wingsState = "ACTIVE";
+        createDuck(runner, "yellow", height, material, "quack", wingsState);
         runner.$(http().client("http://localhost:2222")
                 .receive()
                 .response()
                 .message()
                 .extract(fromBody().expression("$.id", "duckId"))
         );
-        deleteDuck(runner, "${duckId}");
-        validateResponse(runner, "{ \"message\": \"string\" }");
+        duckUpdate(runner, "white", height, Integer.parseInt(context.getVariable("duckId")), material, "quack!!!!!",wingsState);
+        validateResponse(runner, "{ \"message\": \"Duck with id = " + "${duckId}" + " is updated\" }");
     }
 
     public void createDuck(TestCaseRunner runner, String color, double height, String material, String sound, String wingsState) {
@@ -44,12 +47,17 @@ public class DuckDeleteTest extends TestNGCitrusSpringSupport {
                         "\" \n }"));
     }
 
-    public void deleteDuck(TestCaseRunner runner, String id) {
+    public void duckUpdate(TestCaseRunner runner, String color, double height, int id, String material, String sound, String wingsState) {
         runner.$(http().client("http://localhost:2222")
                 .send()
-                .delete("api/duck/delete")
-                .message()
-                .queryParam("id", id));
+                .put("api/duck/update")
+                .queryParam("color", color)
+                .queryParam("height", Double.toString(height))
+                .queryParam("id", Integer.toString(id))
+                .queryParam("material", material)
+                .queryParam("sound", sound)
+                .queryParam("wingsState", wingsState));
+
     }
 
     public void validateResponse(TestCaseRunner runner, String responseMessage) {
